requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==4.4.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.14.0
cachetools==5.5.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
config==0.5.1
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
frozenlist==1.4.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
greenlet==3.0.0
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.4
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.22
starlette==0.26.1
typer==0.12.3
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.2
